♻️refactor: api에 맞춰 변수명 리팩토링
